refactor concurrent and single launches
